backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1

# SQL Server (pymssql is easier to install than pyodbc in Linux/WSL)
pymssql==2.3.1

# Firestore
google-cloud-firestore==2.16.0

# Rate limiting (in-memory, no Redis)
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0  # Required for TestClient
